+++ YandexMobileAdsExample/app/src/androidTest/java/com/yandex/ads/sample/pageobjects/InterstitialScreen.kt
@@ -10,18 +10,6 @@ import com.yandex.ads.sample.components.OrdinalItem
 import io.github.kakaocup.kakao.text.KButton
 import org.hamcrest.Matchers
 
-internal fun InterstitialScreen.clickShowAd() = showAdButton {
-    isVisible()
-    isClickable()
-    click()
-}
-
-internal fun InterstitialScreen.clickCallToAction() = callToActionButton {
-    isVisible()
-    isClickable()
-    click()
-}
-
 internal class InterstitialScreen : KScreen<InterstitialScreen>(),
     HasNetworkMenu<InterstitialScreen.NetworkItem> {
 
@@ -47,18 +35,18 @@ internal class InterstitialScreen : KScreen<InterstitialScreen>(),
         },
         positionProvider = { clazz ->
             when (clazz) {
-                NetworkItem.Vungle::class.java -> 11
-                NetworkItem.UnityAds::class.java -> 10
-                NetworkItem.Tapjoy::class.java -> 9
-                NetworkItem.StartApp::class.java -> 8
-                NetworkItem.Pangle::class.java -> 7
-                NetworkItem.MyTarget::class.java -> 6
-                NetworkItem.Mintegral::class.java -> 5
-                NetworkItem.InMobi::class.java -> 4
-                NetworkItem.Chartboost::class.java -> 3
-                NetworkItem.AppLovin::class.java -> 2
-                NetworkItem.AdMob::class.java -> 1
                 NetworkItem.Yandex::class.java -> 0
+                NetworkItem.AdMob::class.java -> 2
+                NetworkItem.AppLovin::class.java -> 3
+                NetworkItem.Chartboost::class.java -> 6
+                NetworkItem.InMobi::class.java -> 7
+                NetworkItem.Mintegral::class.java -> 8
+                NetworkItem.MyTarget::class.java -> 9
+                NetworkItem.Pangle::class.java -> 11
+                NetworkItem.StartApp::class.java -> 12
+                NetworkItem.Tapjoy::class.java -> 13
+                NetworkItem.UnityAds::class.java -> 14
+                NetworkItem.Vungle::class.java -> 15
                 else -> error("unsupported type")
             }
         }