+++ changelogs/mediation/google/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to AdMob Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 24.9.0.0-beta.1
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.0.0-beta.1
+* Google SDK version 24.9.0
+
 ## Version 23.6.0.10
 
 #### Updated