+++ changelogs/mediation/tapjoy/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to TapJoy Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 14.3.1.9-beta.1
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.0.0-beta.1
+
 ## Version 14.3.1.8
 
 #### Updated